@@ -27,9 +27,9 @@
   ],
   "type": "typo3-cms-extension",
   "require": {
-    "php": ">=7.1 <8.1",
-    "typo3/cms-core": ">=10.4 <11.9.99",
-    "google/cloud-storage": ">=v1.23 <1.27"
+    "php": ">=8.1 <8.3",
+    "typo3/cms-core": ">=12.0 <12.9.99",
+    "google/cloud-storage": ">=v1.30 <1.31"
   },
   "autoload": {
     "psr-4": {
